requirements.txt
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg-binary==3.2.2
psycopg-pool==3.2.2
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vercel_blob==0.2.5.3
waitress==3.0.0
Werkzeug==3.0.3
